[Doc] Fix Gmail rename error + fix gmail_checker neuron
