+++ neurons/__init__.py
@@ -5,7 +5,7 @@ from say import Say
 from script import Script
 from sleep import Sleep
 from systemdate import Systemdate
-from gmail import Gmail
+from gmail_checker import Gmail_checker
 from push_message import Push_message
 from openweathermap import Openweathermap
 from tasker_autoremote import Tasker_autoremote
