--- neurons/gmail/__init__.py
@@ -1,1 +1,0 @@
-from gmail import Gmail
